server/routes/user_route_test.go
package routes

import (
	"bytes"
	"encoding/json"
	"errors"
	"net/http"
	"net/http/httptest"
	"task4/model"
	"testing"
)

// MockUserService is a mock implementation of UserServiceInterface for testing.
type MockUserService struct {
	GetUsersFunc   func() (error, []*model.UserModel)
	AddUserFunc    func(user *model.UserModel) error
	UpdateUserFunc func(user *model.UserModel) error
	DeleteUserFunc func(user *model.UserModel) error
}

func (m *MockUserService) GetUsers() (error, []*model.UserModel) {
	if m.GetUsersFunc != nil {
		return m.GetUsersFunc()
	}
	return nil, []*model.UserModel{}
}

func (m *MockUserService) AddUser(user *model.UserModel) error {
	if m.AddUserFunc != nil {
		return m.AddUserFunc(user)
	}
	return nil
}

func (m *MockUserService) UpdateUser(user *model.UserModel) error {
	if m.UpdateUserFunc != nil {
		return m.UpdateUserFunc(user)
	}
	return nil
}

func (m *MockUserService) DeleteUser(user *model.UserModel) error {
	if m.DeleteUserFunc != nil {
		return m.DeleteUserFunc(user)
	}
	return nil
}

func TestUserRouter_GetUsers(t *testing.T) {
	// Create a mock user service.
	mockUserService := &MockUserService{}

	// Create a UserRouter instance with the mock user service.
	userRouter := NewUserRouter(mockUserService)

	// Create a request.
	req, err := http.NewRequest("GET", "/get-users", nil)
	if err != nil {
		t.Fatal(err)
	}

	// Create a response recorder.
	rr := httptest.NewRecorder()

	// Call the GetUsers method.
	userRouter.GetUsers(rr, req)

	// Check the response status code.
	if rr.Code != http.StatusOK {
		t.Errorf("Expected status code %d, but got %d", http.StatusOK, rr.Code)
	}
}

func TestUserRouter_AddUser(t *testing.T) {
	// Create a mock user service.
	mockUserService := &MockUserService{}

	// Create a UserRouter instance with the mock user service.
	userRouter := NewUserRouter(mockUserService)

	// Create a sample user model.
	user := &model.UserModel{
		// Initialize fields here.
	}

	// Marshal the user model to JSON.
	userJSON, err := json.Marshal(user)
	if err != nil {
		t.Fatal(err)
	}

	// Create a request with the JSON data.
	req, err := http.NewRequest("POST", "/api/add-user", bytes.NewBuffer(userJSON))
	if err != nil {
		t.Fatal(err)
	}
	req.Header.Set("Content-Type", "application/json")

	// Create a response recorder.
	rr := httptest.NewRecorder()

	// Call the AddUser method.
	userRouter.AddUser(rr, req)

	// Check the response status code.
	if rr.Code != http.StatusOK {
		t.Errorf("Expected status code %d, but got %d", http.StatusOK, rr.Code)
	}
}

func TestUserRouter_UpdateUser(t *testing.T) {
	// Create a mock user service.
	mockUserService := &MockUserService{}

	// Create a UserRouter instance with the mock user service.
	userRouter := NewUserRouter(mockUserService)

	// Create a sample user model.
	user := &model.UserModel{
		// Initialize fields here.
	}

	// Marshal the user model to JSON.
	userJSON, err := json.Marshal(user)
	if err != nil {
		t.Fatal(err)
	}

	// Create a request with the JSON data.
	req, err := http.NewRequest("PUT", "/api/update-user", bytes.NewBuffer(userJSON))
	if err != nil {
		t.Fatal(err)
	}
	req.Header.Set("Content-Type", "application/json")

	// Create a response recorder.
	rr := httptest.NewRecorder()

	// Call the UpdateUser method.
	userRouter.UpdateUser(rr, req)

	// Check the response status code.
	if rr.Code != http.StatusOK {
		t.Errorf("Expected status code %d, but got %d", http.StatusOK, rr.Code)
	}
}

func TestUserRouter_DeleteUser(t *testing.T) {
	// Create a mock user service.
	mockUserService := &MockUserService{}

	// Create a UserRouter instance with the mock user service.
	userRouter := NewUserRouter(mockUserService)

	// Create a sample user model.
	user := &model.UserModel{
		// Initialize fields here.
	}

	// Marshal the user model to JSON.
	userJSON, err := json.Marshal(user)
	if err != nil {
		t.Fatal(err)
	}

	// Create a request with the JSON data.
	req, err := http.NewRequest("DELETE", "/api/delete-user", bytes.NewBuffer(userJSON))
	if err != nil {
		t.Fatal(err)
	}
	req.Header.Set("Content-Type", "application/json")

	// Create a response recorder.
	rr := httptest.NewRecorder()

	// Call the DeleteUser method.
	userRouter.DeleteUser(rr, req)

	// Check the response status code.
	if rr.Code != http.StatusOK {
		t.Errorf("Expected status code %d, but got %d", http.StatusOK, rr.Code)
	}
}

func TestUserRouter_GetUsers_Error(t *testing.T) {
	// Create a mock user service that returns an error.
	mockError := errors.New("mock error")

	mockUserService := &MockUserService{
		GetUsersFunc: func() (error, []*model.UserModel) {
			// Define the behavior here.
			return mockError, nil
		},
	}

	// Create a UserRouter instance with the mock user service.
	userRouter := NewUserRouter(mockUserService)

	// Create a request.
	req, err := http.NewRequest("GET", "/get-users", nil)
	if err != nil {
		t.Fatal(err)
	}

	// Create a response recorder.
	rr := httptest.NewRecorder()

	// Call the GetUsers method.
	userRouter.GetUsers(rr, req)

	// Check the response status code.
	if rr.Code != http.StatusInternalServerError {
		t.Errorf("Expected status code %d, but got %d", http.StatusInternalServerError, rr.Code)
	}
}

func TestUserRouter_AddUser_Error(t *testing.T) {
	// Create a mock user service that returns an error.
	mockError := errors.New("mock error")
	// Create a mock user service with custom behavior for AddUser.
	mockUserService := &MockUserService{
		AddUserFunc: func(user *model.UserModel) error {
			// Define the behavior here.
			return mockError
		},
	}

	// Create a UserRouter instance with the mock user service.
	userRouter := NewUserRouter(mockUserService)

	// Create a sample user model.
	user := &model.UserModel{
		// Initialize fields here.
	}

	// Marshal the user model to JSON.
	userJSON, err := json.Marshal(user)
	if err != nil {
		t.Fatal(err)
	}

	// Create a request with the JSON data.
	req, err := http.NewRequest("POST", "/add-user", bytes.NewBuffer(userJSON))
	if err != nil {
		t.Fatal(err)
	}
	req.Header.Set("Content-Type", "application/json")

	// Create a response recorder.
	rr := httptest.NewRecorder()

	// Call the AddUser method.
	userRouter.AddUser(rr, req)

	// Check the response status code.
	if rr.Code != http.StatusInternalServerError {
		t.Errorf("Expected status code %d, but got %d", http.StatusInternalServerError, rr.Code)
	}
}

func TestUserRouter_UpdateUser_Error(t *testing.T) {
	// Create a mock user service that returns an error.
	mockError := errors.New("mock error")
	mockUserService := &MockUserService{
		UpdateUserFunc: func(user *model.UserModel) error {
			// Define the behavior here.
			return mockError
		},
	}

	// Create a UserRouter instance with the mock user service.
	userRouter := NewUserRouter(mockUserService)

	// Create a sample user model.
	user := &model.UserModel{
		// Initialize fields here.
	}

	// Marshal the user model to JSON.
	userJSON, err := json.Marshal(user)
	if err != nil {
		t.Fatal(err)
	}

	// Create a request with the JSON data.
	req, err := http.NewRequest("PUT", "/update-user", bytes.NewBuffer(userJSON))
	if err != nil {
		t.Fatal(err)
	}
	req.Header.Set("Content-Type", "application/json")

	// Create a response recorder.
	rr := httptest.NewRecorder()

	// Call the UpdateUser method.
	userRouter.UpdateUser(rr, req)

	// Check the response status code.
	if rr.Code != http.StatusInternalServerError {
		t.Errorf("Expected status code %d, but got %d", http.StatusInternalServerError, rr.Code)
	}
}

func TestUserRouter_DeleteUser_Error(t *testing.T) {
	mockError := errors.New("mock error")
	// Create a mock user service with custom behavior for DeleteUser.
	mockUserService := &MockUserService{
		DeleteUserFunc: func(user *model.UserModel) error {
			// Define the behavior here.
			return mockError
		},
	}

	// Create a UserRouter instance with the mock user service.
	userRouter := NewUserRouter(mockUserService)

	// Create a sample user model.
	user := &model.UserModel{
		// Initialize fields here.
	}

	// Marshal the user model to JSON.
	userJSON, err := json.Marshal(user)
	if err != nil {
		t.Fatal(err)
	}

	// Create a request with the JSON data.
	req, err := http.NewRequest("DELETE", "/delete-user", bytes.NewBuffer(userJSON))
	if err != nil {
		t.Fatal(err)
	}
	req.Header.Set("Content-Type", "application/json")

	// Create a response recorder.
	rr := httptest.NewRecorder()

	// Call the DeleteUser method.
	userRouter.DeleteUser(rr, req)

	// Check the response status code.
	if rr.Code != http.StatusInternalServerError {
		t.Errorf("Expected status code %d, but got %d", http.StatusInternalServerError, rr.Code)
	}
}

// Add similar error handling tests for other methods (AddUser, UpdateUser, DeleteUser) as needed.
